- Backward compatibility for alert email configuration 6.1.1 and 6.3.1 - Fix for AlertCategoryEnum kArchivalCopy - Added missing enum values. (kAD, kOracleCluster, kInacccessible) - Updated following models   DatastoreInfo:Capacity -> long   DatastoreInfo:FreeSpace -> long   HypervDatastore:Capacity -> long   HypervDatastore:FreeSpace -> long   PhysicalVolume:logicalSizeBytes -> long   PhysicalVolume:usedSizeBytes -> long   SQLProtectionSource:SQLServerDBState -> string   SQLSourceId:InstanceId -> string   ProtectionSourceNode:Nodes -> List<ProtectionSourceNode>   ProtectionSourceNode:ApplicationNodes -> List<ProtectionSourceNode>
- Amended following models
  AwsProtectionSource
  HypervProtectionSource
  VolumeSecurityInfo
  BackupRun
  ProtectionJob
  SnapshotInfo

+++ src/Cohesity.Powershell.Models/DatastoreInfo.cs
@@ -25,7 +25,7 @@ namespace Cohesity.Model
         /// </summary>
         /// <param name="capacity">Specifies the capacity of the datastore in bytes..</param>
         /// <param name="freeSpace">Specifies the available space on the datastore in bytes..</param>
-        public DatastoreInfo(int? capacity = default(int?), int? freeSpace = default(int?))
+        public DatastoreInfo(long? capacity = default(long?), long? freeSpace = default(long?))
         {
             this.Capacity = capacity;
             this.FreeSpace = freeSpace;
@@ -38,14 +38,14 @@ namespace Cohesity.Model
         /// </summary>
         /// <value>Specifies the capacity of the datastore in bytes.</value>
         [DataMember(Name="capacity", EmitDefaultValue=true)]
-        public int? Capacity { get; set; }
+        public long? Capacity { get; set; }
 
         /// <summary>
         /// Specifies the available space on the datastore in bytes.
         /// </summary>
         /// <value>Specifies the available space on the datastore in bytes.</value>
         [DataMember(Name="freeSpace", EmitDefaultValue=true)]
-        public int? FreeSpace { get; set; }
+        public long? FreeSpace { get; set; }
 
         /// <summary>
         /// Returns the string presentation of the object
